Changed image analysis algorithm, how images are accessed in loop etc.

+++ Terminal-executable/test.py
@@ -6,8 +6,8 @@ from nude import *
 from rake import *
 
 if __name__ == "__main__":
-    facebook_id = raw_input("Facebook id linked to Tinder profile: ")
-    facebook_auth_token = raw_input("Facebook authentication token: ")
+    facebook_id = ""
+    facebook_auth_token = ""
     user_limit = input("\nNumber of tests: ")
     image_limit = input("Number of images per test: ")
 
@@ -15,35 +15,37 @@ if __name__ == "__main__":
     total_users = session.nearby_users()
     if len(total_users) > user_limit:
         users = total_users[:user_limit]
-    else:
-        users = total_users
+    else: users = total_users
     
-    user_number = 1
     for user in users:
         print("\n----------\n\nRunning the promiscuity algorithm for " + user.name + ', ' + str(user.age))
         total_skin_percent = 0.0
         bio_score = 0.0
         final_percent = 0.0
 
+
         # Image analysis
-        try:
-            for i in range(image_limit):
-                image_name = str(user_number) + '_' + str(i+1) + '.jpg'
-                photo_url = str(user.get_photos(width='640')[i])
-                print('\nPhoto '+ str(i) + ':\n' + photo_url)
+        image_count = 0
+        for image in user.get_photos(width='320'):
+            if image_count < image_limit:
+                image_count += 1
+                image_name = 'temp'+str(image_count)+'.jpg'
+                photo_url = str(image)        
+                print('\nPhoto '+ str(image_count) + ':\n' + photo_url)
                 image = urllib.URLopener()
                 image.retrieve(photo_url, image_name)
 
                 skin_percent = 100*contains_nudity(image_name)
-                color_skin(image_name)      # only for testing accuracy of nude.py
+                # color_skin(image_name)      # only for testing accuracy of nude.py
                 print('Skin region percentage = ' + str(skin_percent))
                 total_skin_percent += skin_percent
                 os.remove(image_name)
-        except IndexError: print('')
-        
-        total_skin_percent /= image_limit
+            else: break
+            
+        total_skin_percent /= image_count
         print('\nAverage skin region percentage = ' + str(total_skin_percent))
         
+
         # Bio analysis
         try:
             text = user.bio.lower().replace('\n','. ')
@@ -64,6 +66,7 @@ if __name__ == "__main__":
         print(keywords)
         print('Bio score = ' + str(bio_score))
         
+        
         if bio_score != 0:
             final_percent = (total_skin_percent + bio_score)/2
         else:
@@ -71,4 +74,3 @@ if __name__ == "__main__":
         print('\nFinal Score = ' + str(final_percent))
 
         user.like()     # TODO : add conditions to automatic likes
-        user_number += 1
